+++ fe2o3-amqp-types/src/messaging/message/body.rs
@@ -27,7 +27,7 @@ pub enum Body<T> {
     /// the message. However, this is not the way `proton` is implemented, and according to
     /// [PROTON-2574](https://issues.apache.org/jira/browse/PROTON-2574), the wording in the
     /// core specification was an unintended.
-    Nothing,
+    Empty,
 }
 
 impl<T> Body<T> {
@@ -48,7 +48,7 @@ impl<T> Body<T> {
 
     /// Whether the body section is `Nothing`
     pub fn is_nothing(&self) -> bool {
-        matches!(self, Body::Nothing)
+        matches!(self, Body::Empty)
     }
 }
 
@@ -61,7 +61,7 @@ where
             Body::Data(data) => write!(f, "{}", data),
             Body::Sequence(seq) => write!(f, "{}", seq),
             Body::Value(val) => write!(f, "{}", val),
-            Body::Nothing => write!(f, "Nothing"),
+            Body::Empty => write!(f, "Nothing"),
         }
     }
 }
@@ -75,6 +75,12 @@ impl<T: Serialize> From<T> for Body<T> {
 impl<T: Serialize + Clone, const N: usize> From<[T; N]> for Body<T> {
     fn from(values: [T; N]) -> Self {
         Self::Sequence(AmqpSequence(values.to_vec()))
+    }
+}
+
+impl<T> From<AmqpValue<T>> for Body<T> {
+    fn from(value: AmqpValue<T>) -> Self {
+        Self::Value(value)
     }
 }
 
@@ -127,10 +133,10 @@ impl<T: Serialize> Body<T> {
         S: serde::Serializer,
     {
         match self {
-            Body::Data(data) => data.serialize(serializer),
-            Body::Sequence(seq) => seq.serialize(serializer),
-            Body::Value(val) => val.serialize(serializer),
-            Body::Nothing => AmqpValue(()).serialize(serializer),
+            Body::Data(data) => Serializable(data).serialize(serializer),
+            Body::Sequence(seq) => Serializable(seq).serialize(serializer),
+            Body::Value(val) => Serializable(val).serialize(serializer),
+            Body::Empty => Serializable(AmqpValue(())).serialize(serializer),
         }
     }
 }
@@ -207,15 +213,15 @@ where
 
         match val {
             Field::Data => {
-                let data = variant.newtype_variant()?;
+                let Deserializable(data) = variant.newtype_variant()?;
                 Ok(Body::Data(data))
             }
             Field::Sequence => {
-                let sequence = variant.newtype_variant()?;
+                let Deserializable(sequence) = variant.newtype_variant()?;
                 Ok(Body::Sequence(sequence))
             }
             Field::Value => {
-                let value = variant.newtype_variant()?;
+                let Deserializable(value) = variant.newtype_variant()?;
                 Ok(Body::Value(value))
             }
         }