+++ fe2o3-amqp-ext/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.3.0
+
+1. Updated `fe2o3-amqp-types` version to "0.4.0"
+
 ## 0.2.1
 
 1. Added `descriptor_code()` and `descriptor_name` methods